+++ elements/layercontrol/tsconfig.json
@@ -19,5 +19,5 @@
     "checkJs": true,
     "types": ["cypress"]
   },
-  "include": ["src/*.js"]
+  "include": ["src/*.js", "src/helpers/index.js"]
 }
